build.sh: fix build

+++ build.sh
@@ -7,8 +7,13 @@ CROSS_32=arm-linux-gnueabihf-
 rm -f fastboot.bin l-loader.bin
 
 echo "######  BUILD U-BOOT ######"
-cp -f ${UBOOT_CFG} u-boot/.config
-make -C u-boot clean
+if [ -f ${UBOOT_CFG} ]; then
+	cp -f ${UBOOT_CFG} u-boot/.config
+else
+	echo "${UBOOT_CFG} not found!"
+	exit 1
+fi
+make -C u-boot CROSS_COMPILE=${CROSS_64} clean
 make -j4 -C u-boot CROSS_COMPILE=${CROSS_64}
 
 if [ ! -f u-boot/u-boot.bin ]; then
@@ -18,8 +23,8 @@ fi
 
 echo "######  BUILD ATF ######"
 make -C atf distclean
-if [ "$1" = RECOVERY ]; then
-	echo "BUILD ATF WITH DEBUG"
+if [ "$1" = "RECOVERY" ]; then
+	echo "BUILD ATF WITH RECOVERY"
 	make -C atf CROSS_COMPILE=${CROSS_64} all fip \
 		DEBUG=1 PLAT=poplar SPD=none BL33=../u-boot/u-boot.bin \
 		POPLAR_RECOVERY=1
@@ -37,17 +42,18 @@ echo "######  BUILD l-loader ######"
 cp -f ./atf/build/poplar/debug/bl1.bin ./l-loader/atf/
 cp -f ./atf/build/poplar/debug/fip.bin ./l-loader/atf/
 make -C l-loader clean
-if [ "$1" = RECOVERY ]; then
-	echo "BUILD l-loader WITH DEBUG"
-	make -C l-loader CROSS_COMPILE=${CROSS_32} RECOVERY=1
+if [ "$1" = "RECOVERY" ]; then
+	echo "BUILD l-loader WITH RECOVERY"
+	make -C l-loader CROSS_COMPILE=${CROSS_32} \
+		ARM_TRUSTED_FIRMWARE=../atf RECOVERY=1
 else
-	make -C l-loader CROSS_COMPILE=${CROSS_32}
+	make -C l-loader CROSS_COMPILE=${CROSS_32} \
+		ARM_TRUSTED_FIRMWARE=../atf
 fi
 
-if [ "$1" = RECOVERY ]; then
+if [ "$1" = "RECOVERY" ]; then
 	if [ -f l-loader/fastboot.bin ]; then
 		cp -f l-loader/fastboot.bin .
-		echo "######  BUILD DONE ######"
 	else
 		echo "l-loader build failed!"
 		exit 1
@@ -55,10 +61,10 @@ if [ "$1" = RECOVERY ]; then
 else
 	if [ -f l-loader/l-loader.bin ]; then
 		cp -f l-loader/l-loader.bin .
-		echo "######  BUILD DONE ######"
 	else
 		echo "l-loader build failed!"
 		exit 1
 	fi
 fi
+echo "######  BUILD DONE ######"
